+++ app/controllers/search_controller.rb
@@ -6,8 +6,13 @@ class SearchController < ApplicationController
 
   # GET /search
   def index
+    clean_search = params[:search].to_s.squish.downcase
     @member = User.current.find_by("LOWER(username) = ?", params[:search].to_s.downcase)
-    @search = params[:search].to_s.downcase.split(/[^\w]/).reject(&:blank?).uniq.join(" & ")
+    @search = clean_search.split(/[^\w]/).reject(&:blank?).uniq.join(" & ")
     @search_documents = PgSearch.multisearch(params[:search]).page(params[:page]).per(10)
+    if clean_search.present?
+      search = Search.where(search: clean_search).first_or_create
+      search.increment! :search_count
+    end
   end
 end