+++ app/models/search.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+# Track most common searches.
+class Search < ApplicationRecord
+end