+++ db/migrate/20180730171122_create_searches.rb
@@ -1,0 +1,9 @@
+class CreateSearches < ActiveRecord::Migration[5.2]
+  def change
+    create_table :searches do |t|
+      t.text :search
+      t.integer :search_count, null: false, default: 0
+      t.index :search_count
+    end
+  end
+end