+++ test/controllers/admin_controller_test.rb
@@ -91,4 +91,10 @@ class AdminControllerTest < ActionDispatch::IntegrationTest
     end
     assert_redirected_to admin_profile_review_url
   end
+
+  test "should get searches as admin" do
+    login(@admin)
+    get admin_searches_url
+    assert_response :success
+  end
 end